Point to the GitHub Actions logs
If someone were to run the Action multiple times to try and get the number they want, not only would the output files contain duplicate entries for the same day, these runs would also be captured in the logs. There is no way to pull off any shenanigans without leaving a trace.